% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{meanit-package}
\alias{meanit-package}
\title{meanit}
\description{
composite image generation in R
}
\details{
meanit is an R library designed to generate composite PNGs or JPEGs.
In this regard it marks R's rapid approach to the threshold of Zawinski's law.
Prototypes have been used in the generation of \href{http://blog.ironholds.org/art-as-data-as-art-part-ii/}{composite images from open
dataset releases}.
}

