\name{meanit-package}
\docType{package}

\title{
Composite image generation in R
}
\description{
A composite image generator for JPEGS and PNGs.
}
\details{
\tabular{ll}{
Package: \tab meanit\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2014-01-23\cr
License: \tab MIT\cr
}

}
\author{
Oliver Keyes <scire.facias@gmail.com>
}
